refactor /sars-cov-2 & /influenza  components;
Decouple build map from build dropdown
menu. Remove spacers and non-specific
stuff from specific components like
hierarchical dropdown menu and sit-rep
listing. Decouple fetching of build
catalogue from dropdown menu.
